feat(activations): Move ActivateMode to flox-core
So that we can use it in `flox_core::activations`.

Can't wrap `proptest_derive` in `cfg_attr(test)` across crates.
