+++ cli/flox-core/src/activate/mod.rs
@@ -1,2 +1,3 @@
 pub mod context;
+pub mod mode;
 pub mod vars;
