+++ cli/flox-core/src/activate/mode.rs
@@ -1,0 +1,51 @@
+use std::fmt::Display;
+use std::str::FromStr;
+
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(
+    Debug,
+    Clone,
+    Serialize,
+    Deserialize,
+    Hash,
+    PartialEq,
+    Eq,
+    Ord,
+    PartialOrd,
+    Default,
+    JsonSchema,
+    proptest_derive::Arbitrary,
+)]
+#[serde(rename_all = "kebab-case")]
+pub enum ActivateMode {
+    #[default]
+    Dev,
+    Run,
+}
+
+impl Display for ActivateMode {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ActivateMode::Dev => write!(f, "dev"),
+            ActivateMode::Run => write!(f, "run"),
+        }
+    }
+}
+
+#[derive(Debug, thiserror::Error)]
+#[error("not a valid activation mode")]
+pub struct ActivateModeParseError;
+
+impl FromStr for ActivateMode {
+    type Err = ActivateModeParseError;
+
+    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
+        match s {
+            "dev" => Ok(ActivateMode::Dev),
+            "run" => Ok(ActivateMode::Run),
+            _ => Err(ActivateModeParseError),
+        }
+    }
+}
