+++ cli/flox/src/commands/services/mod.rs
@@ -2,13 +2,13 @@ use std::path::{Path, PathBuf};
 
 use anyhow::Result;
 use bpaf::Bpaf;
-use flox_core::activate::context::InvocationType;
+use flox_core::activate::context::{ActivateMode, InvocationType};
 use flox_rust_sdk::data::System;
 use flox_rust_sdk::flox::Flox;
 use flox_rust_sdk::models::environment::Environment;
 use flox_rust_sdk::models::environment::generations::GenerationId;
 use flox_rust_sdk::models::lockfile::Lockfile;
-use flox_rust_sdk::models::manifest::typed::{ActivateMode, Inner, Manifest, Services};
+use flox_rust_sdk::models::manifest::typed::{Inner, Manifest, Services};
 use flox_rust_sdk::providers::services::process_compose::{
     ProcessState,
     ProcessStates,
