+++ cli/flox/src/utils/active_environments.rs
@@ -4,10 +4,10 @@ use std::fmt::{self, Display};
 use std::str::FromStr;
 
 use anyhow::Result;
+use flox_core::activate::mode::ActivateMode;
 use flox_core::activate::vars::FLOX_ACTIVE_ENVIRONMENTS_VAR;
 use flox_rust_sdk::models::environment::UninitializedEnvironment;
 use flox_rust_sdk::models::environment::generations::GenerationId;
-use flox_rust_sdk::models::manifest::typed::ActivateMode;
 use serde::{Deserialize, Serialize};
 use tracing::debug;
